optimized PrimeSieve class for space, solved 203 and 187

+++ euler/007 10001st prime.cpp
@@ -2,7 +2,7 @@
 #include "PrimeSieve.hpp"
 
 int main(){
-	pe::PrimeSieve<int> s(110000);
+	pe::PrimeSieve<long> s(110000);
 	std::cout << s.nth_prime(10001) << std::endl;
 	return 0;
 }
